made dependency analysis for domain axioms more fine grained
